requirements.txt
python-digitalocean
flask
requests
werkzeug
flask_cors
colorlog
flask-restful

